+++ app/routes/ffxiv.craftsim.tsx
@@ -35,7 +35,7 @@ import {
   parseStringToNumberArray,
   parseZodErrorsToDisplayString
 } from '~/utils/zodHelpers'
-import DoHFilter from '~/components/form/ffxiv/DoHFilter'
+import Filter from '~/components/form/Filter'
 import {
   getActionUrl,
   handleCopyButton,
@@ -224,7 +224,10 @@ export default function Index() {
               type="button"
             />
           </div>
-          <DoHFilter
+          <Filter
+            formName="jobs"
+            filterButtonText="Choose DoH"
+            selectedCountText="DoH"
             defaultValue={loaderData.jobs}
             options={dOHOptions}
             title={inputMap.jobs}
